requirements.txt
loguru==0.6.0
numpy==1.24.1
pytz==2022.7.1
websocket-client==1.5.0
